Room control & variable spawn points
apologies for making changes to scene_level_0, didn't mean to do that.

Currently what is going on is that MatchManager registers all of the
rooms with a Control script and counts them as spawn points, as well as
a default one for each team. When rooms get shuffled around, they switch
from list to list. Crim spawn points are sorted in order of X position.
A specific "cloning" room could be made that does this kind of thing
exclusively.

Furthermore, rooms now will revert to their faction when left alone
mid-hack. Cops can reclaim rooms, and having a contested room will
result in a stalemate of the hacking counter.

Also updated SpawnWeapons to handle the fact that faction changes can
happen both ways.