eliminated friendly fire via collision layers
cops, cop robots and their projectiles go on the "Cops" layer,
likewise for crims.

Things on those respective layers won't collide with things on the same
layer. See Edit->Project Settings -> Physics for the collision matrix.

Added the Utils class for generally useful stuff. In this case, layers
are only accessible by their INDEX VALUE, so there are convenience
functions to return the layer number. That way there's a single point of
maintenance if / when the layers inevitably get fucked with.