added clone room, example of destructible stations
clone room script is pretty bare-bones, but it's an example of the kind
of things to take care of to have rooms register whichever thing can
take damage, and matchManager asks a clone room if it is broken or not
before allowing spawning to happen

It's probably high time to draw up a Room base-class that takes care of
a lot of this sort of bullshit so that everyone doesn't need to worry
about it too much.

See Scene_level_0 for the clone room - you can destroy and repair the
computer console in there, but you can't damage it once you control the
room. neat!