another robot parameter
timer that governs how long a robot will pause after shooting. The value
should be less than the overall time it takes for a robot to be ready to
shoot again, so that the robot will take a shot, wait a bit, move some
more, and then be able to shoot once again.