OKAY SHOOTING STUFF PLZ READ
okay, now the pistol works on trigger press, so does the shotgun. you
can spam the pistol to fire as fast as you can pull the trigger, but not
so with the shotgun. with the smg, you can hold down the trigger to fire
as often as the timer allows