fixed issue with cap sound not turning off for...
...barrier rooms when captured.

Also messing around with what is likely to become balancing hell. Having
an smg in the first area is probably pretty OP. No way to know without
real tests, but I've severely reduced the total ammo that you can have
with the smg, and increased robot damage to 2 instead of 1.

Many of these things are now editable in the inspector so it'll be
easier to tweak them.